requirements.txt
python-dotenv==0.19.0
machine==0.0.1
gpiozero==1.6.2
Adafruit-PCA9685==1.0.1
Adafruit-Blinka==6.13.0
picamera==1.13
adafruit-circuitpython-servokit==1.3.4 
websockets==10.0
redis==3.5.3
smbus==1.1.post2
